events: P1 reliability fixes for scale readiness
P1-1: event consumer drains with BRPOPLPUSH into an in-flight processing list
(was BRPOP = at-most-once); success LREMs, handled failures dead-letter to a
bounded list, and startup reclaims orphans left by a crash. Re-apply is safe
via the idempotent completion ledger — no more silently-dropped credit on a
restart mid-apply.
P1-6: events:active gate carries a 180s TTL refreshed by the consumer, so a
dead consumer stops producers enqueuing instead of LPUSHing forever (Redis OOM).
P1-8: landing a new tile clears any banked choose_task candidates, closing the
draw-easy / roll-onto-hard / apply-banked-easy exploit.
P1-12: populate-random is hard-capped (2000/call) so it can't fire tens of
thousands of statements in one transaction.
P1-7: EventProgress.progress + EventCompletion.quantity are BigInteger (loot_value
folds raw GP → would overflow signed INT ~2.1B and poison later drops); live
schema ALTER in scripts/sql/p1_7_event_progress_bigint.sql for the deploy window.

+++ db/models/events.py
@@ -474,7 +474,9 @@ class EventCompletion(Base):
     team_id = Column(Integer, ForeignKey("web_event_teams.id"), nullable=True)
     player_id = Column(Integer, ForeignKey("players.player_id"), nullable=True)
     status = Column(String(16), nullable=False)  # EVENT_COMPLETION_STATUSES
-    quantity = Column(Integer, nullable=False, default=1)
+    # BigInteger (P1-7): loot_value tasks fold raw GP here — a large multi-day
+    # event can exceed signed-INT 2.147B and poison every later matching drop.
+    quantity = Column(BigInteger, nullable=False, default=1)
     source_type = Column(String(16), nullable=True)  # drop|pb|clog|ca|experience|manual|bonus
     source_id = Column(BigInteger, nullable=True)
     submission_guid = Column(String(64), nullable=True)
@@ -504,7 +506,8 @@ class EventProgress(Base):
     event_id = Column(Integer, ForeignKey("web_events.id"), nullable=False)
     task_id = Column(Integer, ForeignKey("web_event_tasks.id"), nullable=False)
     team_id = Column(Integer, ForeignKey("web_event_teams.id"), nullable=False)
-    progress = Column(Integer, nullable=False, default=0)
+    # BigInteger (P1-7): loot_value progress folds raw GP; see EventCompletion.
+    progress = Column(BigInteger, nullable=False, default=0)
     completed = Column(Boolean, nullable=False, default=False)
     completed_at = Column(DateTime, nullable=True)
